requirements.txt
# Core Discord
discord.py[voice]>=2.3.0
py-cord>=2.4.0

# Web Framework
flask>=2.3.0
flask-session>=0.4.0
flask-sqlalchemy>=3.0.0
flask-limiter>=3.0.0
flask-cors>=4.0.0

# Database
motor>=3.3.2
pymongo>=4.5.0
redis>=4.5.0

# Security
bcrypt>=4.0.0
cryptography>=40.0.0
pyjwt>=2.7.0

# Music & Audio
yt-dlp>=2023.7.0
pynacl>=1.5.0

# Utilities
aiohttp>=3.8.0
requests>=2.31.0
python-dotenv>=1.0.0
pillow>=10.0.0

# AI & ML
openai>=0.27.0

# Testing
pytest>=7.3.0
pytest-asyncio>=0.21.0